+++ Application/Purchases/PurchaseAbstract.cs
@@ -6,6 +6,6 @@ namespace Application.Purchases
     {
         public Guid CustomerId { get; set; }
 		public Guid ProductId { get; set; }
-		public int QuantityPurchased { get; set; }
+		public decimal QuantityPurchased { get; set; }
     }
 }
