+++ Domain/Purchase.cs
@@ -5,7 +5,7 @@ namespace Domain
 		public Guid ProductId { get; set; }
 		public Guid OrderId { get; set; }
 		public Product Product { get; set; }
-		public int QuantityPurchased { get; set; }
+		public decimal QuantityPurchased { get; set; }
 		public DateTime DatePurchased { get; set; } = DateTime.UtcNow;
 		public Order Order {get; set; }
 		public decimal? Discount { get; set; }
